Move onboarding screen ViewModels into their own files
Android Studio does this with a few clicks via Refactor (F6). Moved
them next to the PermissionsViewModel.
