+++ app/src/main/java/org/scottishtecharmy/soundscape/screens/onboarding/Hearing.kt
@@ -32,28 +32,11 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
 import org.scottishtecharmy.soundscape.R
-import org.scottishtecharmy.soundscape.audio.NativeAudioEngine
 import org.scottishtecharmy.soundscape.components.OnboardButton
 import org.scottishtecharmy.soundscape.ui.theme.IntroTypography
 import org.scottishtecharmy.soundscape.ui.theme.IntroductionTheme
-import javax.inject.Inject
-
-@HiltViewModel
-class HearingViewModel @Inject constructor(private val audioEngine : NativeAudioEngine): ViewModel() {
-
-    fun playSpeech(speechText: String) {
-        // Set our listener position, and play the speech
-        // TODO: If updateGeometry isn't called, then the audioEngine doesn't move on to   the next
-        //  queued text to speech. That resulted in the Listen button only working one time.
-        //  Calling updateGeometry (which in the service is called every 30ms) sorts this out.
-        //  We should consider another way of doing this.
-        audioEngine.updateGeometry(0.0, 0.0,0.0)
-        audioEngine.createTextToSpeech(0.0,0.0, speechText)
-    }
-}
+import org.scottishtecharmy.soundscape.viewmodels.HearingViewModel
 
 @Composable
 private fun LocalImage() {
